% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_available_tracks.R
\name{add_available_tracks}
\alias{add_available_tracks}
\title{Calculate the summary of the available tracks}
\usage{
add_available_tracks(local, remote)
}
\arguments{
\item{local}{the connection to the local database}

\item{remote}{the connection to the remote database}
}
\description{
The function calculates the number of tracks per combination of scheme,
classification and species.
}
