% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vleemo_observed_track.R
\name{add_vleemo_observed_track}
\alias{add_vleemo_observed_track}
\title{Import basic track information}
\usage{
add_vleemo_observed_track(local, remote, scheme = "m202206")
}
\arguments{
\item{local}{the connection to the local database}

\item{remote}{the connection to the remote database}

\item{scheme}{the scheme to import}
}
\description{
Select all tracks with manual observations.
The minimal duration is 1 minute.
The airspeed is between 5 and 30 m/s.
}
