% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vleemo_track.R
\name{add_vleemo_track}
\alias{add_vleemo_track}
\title{Import basic track information}
\usage{
add_vleemo_track(
  local,
  remote,
  scheme = "m202206",
  classification = c("MEDIUM_BIRD", "SMALL_BIRD", "LARGE_BIRD")
)
}
\arguments{
\item{local}{the connection to the local database}

\item{remote}{the connection to the remote database}

\item{scheme}{the scheme to import}

\item{classification}{Select only unobserved tracks with this classification.}
}
\description{
Select all tracks without manual observations.
The minimal duration is 1 minute.
The airspeed is between 5 and 30 m/s.
}
