% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vleemo_track_points.R
\name{add_vleemo_track_points}
\alias{add_vleemo_track_points}
\title{Import raw tracks and convert them to equal time spaced tracks}
\usage{
add_vleemo_track_points(local, remote)
}
\arguments{
\item{local}{the connection to the local database}

\item{remote}{the connection to the remote database}
}
\description{
Import raw tracks and convert them to equal time spaced tracks
}
