% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_wind.R
\name{add_wind}
\alias{add_wind}
\title{Calculate the average wind speed and direction for tracks}
\usage{
add_wind(local, remote)
}
\arguments{
\item{local}{the connection to the local database}

\item{remote}{the connection to the remote database}
}
\description{
Calculate the average wind speed and direction for tracks
}
