% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{adjustExprLevel,MeRIP.RADAR-method}
\alias{adjustExprLevel,MeRIP.RADAR-method}
\title{adjustExprLevel}
\usage{
\S4method{adjustExprLevel}{MeRIP.RADAR}(object, adjustBy = "geneSum")
}
\arguments{
\item{object}{The MeRIP.RADAR object that has been normalized for library size.}

\item{adjustBy}{By default, adjust post-IP count by INPUT geneSum. Can also choose "pos" to use current position count to adjust for expression level.}
}
\value{
object The MeRIP.RADAR object now with IP-count adjusted for expression level.
}
\description{
adjustExprLevel
}
