% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustExprLevel.R
\name{adjustExprLevel}
\alias{adjustExprLevel}
\title{adjustExprLevel}
\usage{
adjustExprLevel(x, adjustBy = "geneSum")
}
\arguments{
\item{x}{The data list generated by normlizeLibrary() function.}

\item{adjustBy}{By default, adjust post-IP count by INPUT geneSum. Can also choose "pos" to use current position count to adjust for expression level.}
}
