\name{ApertureWeightingFunctionsAntenna}
\alias{ApertureWeightingFunctionsAntenna}
\title{Antenna Characteristics for Aperture Weighting Functions
}
\usage{ApertureWeightingFunctionsAntenna}
\description{\code{ApertureWeightingFunctionsAntenna} has Antenna Characteristics for 
Aperture Weighting Functions
}
\source{
G. Richard Curry, 2011
SciTech Publishing
Radar Essentials, A Concise Handbook for Radar Design and Performance Analysis
}
\references{
G. Richard Curry, 2011
SciTech Publishing
Radar Essentials, A Concise Handbook for Radar Design and Performance Analysis
}
\author{Jose Gama}
\examples{
data(ApertureWeightingFunctionsAntenna)
str(ApertureWeightingFunctionsAntenna)
}
\keyword{datasets}
