% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beam_block.R
\name{beam_block}
\alias{beam_block}
\title{Calculates the percentage of beam blockage across a radar range}
\usage{
beam_block(data, radar, angle = 0.5,
  nexrad_n = "V:/Documents/nexrad_site_list_with_utm.csv")
}
\arguments{
\item{data}{data.frame with numberic columns for range, azimuth, and ground height (groundht)}

\item{radar}{the 3 letter name  of the radar}

\item{angle}{the angle of the beam can either be a scalar or vector}

\item{nexrad_n}{location of the nexrad site table. Default is the location on the AP's server}
}
\description{
This function requires input of a dataframe with three columns 'range','azimuth', and 'groundht'
}

