% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamht_radio.R
\name{beamht_radio}
\alias{beamht_radio}
\title{Calculates the height of the radar beam using radiosondes}
\usage{
beamht_radio(radio_n, radar, range, total_elev,
  nexrad_n = "V:/Documents/nexrad_site_list_with_utm.csv")
}
\arguments{
\item{radio_n}{file path of the radiosonde in .html}

\item{radar}{the 3 letter radar code}

\item{range}{numeric range of each pulse volume}

\item{total_elev}{a vector of the unique elevations you want to calculate beam heights for}

\item{nexrad_n}{location of the nexrad site table. Default is the location on the AP's server}
}
\description{
Output is a list containing the bottom, middle, and top of the beam.  
Use beamht_std if you want to calculate beam heights using standard refraction
}

