% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamht_std.R
\name{beamht_std}
\alias{beamht_std}
\title{Calculates the height of the radar beam using standard refraction}
\usage{
beamht_std(range, groundht = 0, elev = 0.5, radar = radar,
  nexrad_n = "V:/Documents/nexrad_site_list_with_utm.csv")
}
\arguments{
\item{range}{numeric range of each pulse volume}

\item{groundht}{the ground height above sea level at each pulse volume}

\item{elev}{the elevation of  the radar beam. Can be a scalar or vector}

\item{radar}{the 3 letter radar code}

\item{nexrad_n}{location of the nexrad site table. Default is the location on the AP's server}
}
\description{
Output is a list containing the bottom, middle, and top of the beam.  
Use beamht_radio if you want to calculate beam heights using radiosonde
}

