% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{calc_concordance_index}
\alias{calc_concordance_index}
\title{Compute concordance index for radiomic features}
\usage{
calc_concordance_index(rdr = NULL, surv_obj = NULL,
  thr_concordance = NA, which_data = "scaled")
}
\arguments{
\item{rdr}{A RadAR object (class \code{\link{SummarizedExperiment}}).}

\item{surv_obj}{An object of class \code{\link{Surv}}.}

\item{thr_concordance}{(numeric) Threshold to identify statistically significant features from concordance analysis.
It should be in the range (0, .5].}

\item{which_data}{(character) Which data use to compute concordance index. It can be one of the following: "normal",
"scaled", "normalized".}
}
\value{
An updated rdr (a RadAR object)
}
\description{
This function calculates concordance index of radiomic features.
Response should be a survival object obtained by \code{\link{Surv}}
}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
