% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{calc_cox_regression}
\alias{calc_cox_regression}
\title{Fit a Cox regression model using radiomic features}
\usage{
calc_cox_regression(rdr = NULL, surv_obj = NULL, thr_cox_zvalue = NA,
  which_data = "scaled", multiple_regression = F)
}
\arguments{
\item{rdr}{A RadAR object (class \code{\link{SummarizedExperiment}}).}

\item{surv_obj}{An object of class \code{\link{Surv}}.}

\item{thr_cox_zvalue}{(numeric) Z threshold (Wald statistics) to identify statistically significant features from cox regression analysis.
It should be in the range (0, inf).}

\item{which_data}{(character) Which data use to compute concordance index. It can be one of the following: "normal",
"scaled", "normalized".}

\item{multiple_regression}{(character). Cox regression can be "multiple" (surv ~ all features) or "single" (surv ~ each feature).}
}
\value{
An updated rdr (a RadAR object)
}
\description{
This function fits a Cox Proportional-Hazards Model using radiomic features as independent variables.
Response should be an object obtained by \code{\link{Surv}}
}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
