% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{calc_differential_radiomics}
\alias{calc_differential_radiomics}
\title{Perform differential radiomic analysis}
\usage{
calc_differential_radiomics(rdr = NULL, conditions = NULL,
  which_data = "scaled", method = "wilcox", adjust_pvalues_by = "BH",
  thr_pvalue = NA, thr_auc = NA)
}
\arguments{
\item{rdr}{A RadAR object (class \code{\link{SummarizedExperiment}}).}

\item{conditions}{(numeric, character) Vector of labels for each sample in rdr. Should have the same
length of \code{ncol(rdr)}. Required.}

\item{which_data}{(character) Which data use for plot. It can be one of the following: "normal", "scaled", "normalized".}

\item{method}{(character) Which statistical methods use for differential analysis.
It can be "wilcox" (Wilcoxon-Mann-Whitney), "auc" (Area Under the Curve) or "kruskal-wallis".}

\item{adjust_pvalues_by}{(character) Which method use to correct p-values from wilcox test. Print available methods by
\code{\link{p.adjust.methods}}.}

\item{thr_pvalue}{(numeric) P-value threshold to identify statistically significant features for wilcox or kruskal-wallis.
It should be in the range (0, 1].}

\item{thr_auc}{(numeric) AUC threshold to identify statistically significant features.
It should be in the range (0.5,1].}
}
\value{
An updated rdr (a RadAR object)
}
\description{
This function implements two non-parametric statistical tests (Wilcoxon-Mann-Whitney and AUC)
to perform differential analysis of radiomic features. In case of Wilcoxon-Mann-Whitney p-values can be adjusted using
different methods.
}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
