% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{clockz}
\alias{clockz}
\title{Adds a clock to a for loop}
\usage{
clockz(x, var)
}
\arguments{
\item{x}{the vector used in the for loop}

\item{var}{what is the variable name in the for loop}
}
\description{
Adds a clock to a for loop
}

