\name{ConversiondBZ2Z}
\alias{ConversiondBZ2Z}
\title{Conversion from dBZ (log) units to linear Z units}
\usage{ConversiondBZ2Z(dBZ)}
\description{\code{ConversiondBZ2Z} Converts from dBZ (log) units
to linear Z units
}
\arguments{
  \item{dBZ}{ logarithmic reflectivity value}
}
\value{
  \item{Z}{linear reflectivity units}
}
\source{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}
}
\references{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}
}
\author{Jose Gama}
\keyword{programming}
