\name{ConversionZ2dBZ}
\alias{ConversionZ2dBZ}
\title{Conversion from linear Z units to dBZ (log) units}
\usage{ConversionZ2dBZ(Zlin)}
\description{\code{ConversionZ2dBZ} Converts from linear Z units
to dBZ (log) units
}
\arguments{
  \item{Zlin}{linear reflectivity units}
}
\value{
  \item{dBZ}{ logarithmic reflectivity value}
}
\source{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}
}
\references{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}
}
\author{Jose Gama}
\keyword{programming}
