% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countIntronReads.R
\name{countIntronReads}
\alias{countIntronReads}
\title{countIntronReads}
\usage{
countIntronReads(samplenames, gtf, fragmentLength = 150, bamFolder,
  strandToKeep = "opposite", paired = FALSE, threads = 1)
}
\arguments{
\item{bamFolder}{Path to the folder where bam file locates}

\item{strandToKeep}{According to library preparation protocol, choose which strand to count. Stranded RNA library usually seq the "ooposite" strand. Small RNA library seq the "same" strand.}

\item{paired}{Logical indicating whether the input bam files are from paired end sequencing. Default is FALSE. If using paired end data, the read length will be estimated from the data and only good mate are counted.}

\item{threads}{The number of threads to use for hyperthreading}

\item{binSize}{The size of consecutive bins to slice the transcripts}

\item{outputDir}{The directory to save output files}
}
