% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countReads.R
\name{countReads}
\alias{countReads}
\title{countReads}
\usage{
countReads(samplenames, gtf, shift = 75, bamFolder, outputDir = NA,
  modification = "m6A", binSize = 50, strandToKeep = "opposite",
  threads = 1)
}
\arguments{
\item{samplenames}{The names of each sample (prefix for bam files)}

\item{gtf}{The gtf format gene annotation file}

\item{shift}{Number of bp to shift when counting the reads. This is usually half of the fragment length.}

\item{bamFolder}{Path to the folder where bam file locates}

\item{outputDir}{The directory to save output files}

\item{binSize}{The size of consecutive bins to slice the transcripts}

\item{strandToKeep}{According to library preparation protocol, choose which strand to count. Stranded RNA library usually seq the "ooposite" strand. Small RNA library seq the "same" strand.}

\item{threads}{The number of threads to use for hyperthreading}
}
