% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffIP.R
\name{diffIP}
\alias{diffIP}
\title{diffIP}
\usage{
diffIP(x, Covariates = NULL, plotPvalue = TRUE, p.adjBy = "fdr")
}
\arguments{
\item{x}{The RNADMethyl data list}

\item{Covariates}{Matrix of covariates to include in the PoissonGamma test. Default is NULL.}

\item{p.adjBy}{The method used to adjust for the p value. Default is 'fdr'. Can be "bonferroni","BY","fdr".}
}
