% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{diffIP_parallel}
\alias{diffIP_parallel}
\title{Perform inferential test using Poisson Random effect model in RADAR}
\usage{
\S4method{diffIP_parallel}{MeRIP.RADAR}(object, exclude = NULL,
  maxPsi = 100, fdrBy = "qvalue", thread = 8)
}
\arguments{
\item{object}{The MeRIP.RADAR object}

\item{exclude}{A vector of characters. The samples to be excluded from the differential test.}

\item{maxPsi}{The max random effect parameter Psi allowed in the model fitting.}

\item{fdrBy}{The method to control for false discovery rate. The default is "qvalue", can also be "fdr".}

\item{thread}{The number of thread to use for computing.}
}
\description{
Perform inferential test using Poisson Random effect model in RADAR
}
