% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{do_hierarchical_clustering}
\alias{do_hierarchical_clustering}
\title{Perform hierarchical clustering of the radiomic dataset}
\usage{
do_hierarchical_clustering(rdr = NULL, which_data = "scaled",
  method_dist_row = "euclidean",
  method_dist_col = "correlation.pearson", method_hcl_row = "ward.D",
  method_hcl_col = "ward.D")
}
\arguments{
\item{rdr}{A RadAR object (class \code{\link{SummarizedExperiment}}).}

\item{which_data}{(character) Which data use for the computation of the correlation coefficients. It can be one of the following: "normal",
"scaled", "normalized".}

\item{method_dist_row}{(character) Which method use to computed distance matrix between features (rows). Print available
methods by \code{\link{print_distance_methods}}}

\item{method_dist_col}{(character) Which method use to computed distance matrix between samples (columns). Print available
methods by \code{\link{print_distance_methods}}}

\item{method_hcl_row}{(character) Which method use to generate hierarchical clustering for features (rows). Print available
methods by \code{\link{print_hcl_methods}}}

\item{method_hcl_col}{(character) Which method use to generate hierarchical clustering for samples (columns). Print available
methods by \code{\link{print_hcl_methods}}}
}
\value{
An updated rdr (a RadAR object)
}
\description{
Perform hierarchical clustering of the radiomic dataset
}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
