\name{DopplerDilemma}
\alias{DopplerDilemma}
\title{Doppler dilemma}
\usage{DopplerDilemma(inFloat, lam, speedOfLight)}
\description{\code{DopplerDilemma} returns the Doppler dilemma
From Rinehart (1997), Eqn 6.12
}
\arguments{
  \item{inFloat}{Nyquist Velocity [m/s] or Maximum unambiguous range [m]}
  \item{lam}{Radar wavelength [m]}
  \item{speedOfLight}{speed of light}
}
\value{
  \item{Rmax}{ Maximum unambiguous range [m]}
}
\source{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}

R. E. Rinehart, 1997
Radar for Meteorologists
Rinehart Publishing
}
\references{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}

R. E. Rinehart, 1997
Radar for Meteorologists
Rinehart Publishing
}
\author{Jose Gama}
\keyword{programming}
