\name{DopplerFmax}
\alias{DopplerFmax}
\title{Maximum frequency given PRF}
\usage{DopplerFmax(PRF)}
\description{\code{DopplerFmax} returns the PRF for a  maximum frequency
From Rinehart (1997), Eqn 6.8
}
\arguments{
  \item{PRF}{Pulse repetition frequency [Hz]}
}
\value{
  \item{f}{ Maximum frequency [Hz]}
}
\source{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}

R. E. Rinehart, 1997
Radar for Meteorologists
Rinehart Publishing
}
\references{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}

R. E. Rinehart, 1997
Radar for Meteorologists
Rinehart Publishing
}
\author{Jose Gama}
\keyword{programming}
