\name{DopplerPulseDuration}
\alias{DopplerPulseDuration}
\title{Pulse duration from pulse length}
\usage{DopplerPulseDuration(tau, speedOfLight)}
\description{\code{DopplerPulseDuration} Converts from pulse length to pulse duration 
}
\arguments{
  \item{tau}{ Pulse length [m]}
  \item{speedOfLight}{speed of light}
}
\value{
  \item{pDur}{Pulse duration [s]}
}
\source{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}
}
\references{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}
}
\author{Jose Gama}
\keyword{programming}
