\name{DopplerPulseLength}
\alias{DopplerPulseLength}
\title{Pulse length from pulse duration}
\usage{DopplerPulseLength(pDur, speedOfLight)}
\description{\code{DopplerPulseLength} Converts from pulse duration to pulse length 
}
\arguments{
  \item{pDur}{Pulse duration [s]}
  \item{speedOfLight}{speed of light}
}
\value{
  \item{tau}{ Pulse length [m]}
}
\source{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}
}
\references{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}
}
\author{Jose Gama}
\keyword{programming}
