\name{DopplerRmax}
\alias{DopplerRmax}
\title{Maximum unamiguous range}
\usage{DopplerRmax(PRF, speedOfLight)}
\description{\code{DopplerRmax} returns the maximum unamiguous range
From Rinehart (1997), Eqn 6.11
}
\arguments{
  \item{PRF}{Pulse repetition frequency [Hz]}
  \item{speedOfLight}{speed of light}
}
\value{
  \item{Rmax}{ Maximum unambiguous range [m]}
}
\source{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}

R. E. Rinehart, 1997
Radar for Meteorologists
Rinehart Publishing
}
\references{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}

R. E. Rinehart, 1997
Radar for Meteorologists
Rinehart Publishing
}
\author{Jose Gama}
\keyword{programming}
