\name{DopplerWavelength}
\alias{DopplerWavelength}
\title{Wavelength given frequency}
\usage{DopplerWavelength(freq, speedOfLight)}
\description{\code{DopplerWavelength} Converts from frequency to wavelength 
}
\arguments{
  \item{freq}{ Frequency [Hz]}
  \item{speedOfLight}{speed of light}
}
\value{
  \item{lam}{Wavelength [m]}
}
\source{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}
}
\references{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}
}
\author{Jose Gama}
\keyword{programming}
