% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{dtime_2_clock}
\alias{dtime_2_clock}
\title{Converts decimal time to an as.POSIX time.}
\usage{
dtime_2_clock(date, d_time)
}
\arguments{
\item{date}{the date in the format : yyyy-mm-dd}

\item{d_time}{decimal time as a numeric. Can be in 24+ time however it will not change the date accordingly}
}
\description{
The output has a generic tz of UTC.
}

