% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_methods.R
\name{edgeR}
\alias{edgeR}
\title{edgeR}
\usage{
edgeR(countdata, pheno, covariates = NULL, normalizeLib = FALSE)
}
\arguments{
\item{countdata}{Matrix of read count.}

\item{pheno}{Matrix defining the grouping of the samples.}

\item{covariates}{Optional matrix defining the covariates.}

\item{normalizeLib}{Logic option to choose whether perform normalization. Default if "FALSE" for normalized count matrix.}
}
\description{
A wrapper function to call edgeR
}
