\name{ElectronicWarfareFrequencyBands}
\alias{ElectronicWarfareFrequencyBands}
\title{Electronic Warfare Frequency Bands}
\usage{ElectronicWarfareFrequencyBands}
\description{\code{ElectronicWarfareFrequencyBands} has Electronic Warfare Frequency Bands
}
\source{
G. Richard Curry, 2011
SciTech Publishing
Radar Essentials, A Concise Handbook for Radar Design and Performance Analysis
}
\references{
G. Richard Curry, 2011
SciTech Publishing
Radar Essentials, A Concise Handbook for Radar Design and Performance Analysis
}
\author{Jose Gama}
\examples{
data(ElectronicWarfareFrequencyBands)
str(ElectronicWarfareFrequencyBands)
}
\keyword{datasets}
