% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_time_track.R
\name{equal_time_track}
\alias{equal_time_track}
\title{Convert a irregular track into a track with equal time intervals}
\usage{
equal_time_track(raw_track)
}
\arguments{
\item{raw_track}{A single track as a data.frame with projected space-time
coordinates.
Coordinates should be in meters, the time in seconds.}
}
\description{
The resulting track has
}
\details{
\itemize{
\item the part of the track (\code{part}).
We split the part at the points where the ground speed is above 30 meter
per second.
\item the space-time coordinates (\code{t}, \code{x}, \code{y} and \code{z})
}
}
