% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exodus_copy.R
\name{exodus_copy}
\alias{exodus_copy}
\title{Transfers raw radar files from a raw folder to an output folder based on a moving range of sunangles}
\usage{
exodus_copy(birdnight, rawfolder, output, radar, langle, eangle, morning = F,
  years = unique(nights2$year) + 1900, folders = T,
  nexrad_n = "V:/Documents/nexrad_site_list_with_utm.csv")
}
\arguments{
\item{birdnight}{nights you want to select sweeps. Date needs to be in the format: 'mm/dd/yyyy'}

\item{rawfolder}{base folder where raw files are located. Terminal folder must be years}

\item{output}{location of the selected files}

\item{radar}{3 letter radar code}

\item{langle}{late angle. For evenings this is the larger angle.  Mornings is vice/versa}

\item{eangle}{early angle. For evenings this is the smaller angle. Mornings is vice/versa}

\item{morning}{T/F whether you want to grab morning times. This reverses the sun angles and only grabs files from the morning}

\item{years}{a vector of the unique years you want to select from. Default is all years in the birdnight vector}

\item{folders}{Whether you want each night to be seperated into seperate folders. This is required if you're using these files to calculate exodus curves.}

\item{nexrad_n}{location of the nexrad site table. Default is the location on the AP's server}
}
\description{
This program is specifically for grabbing sweeps within a defined window around a specific angle 
File folders for raw folders must be split into seperate year folders. The program is expecting the file system to be: rawfolder/2010 ,rawfolder/2011, rawfolder/2012. If you have these seperated differently, you need to change the file system so that the terminal folder is the year. If necessary you can input a complex path in the output field. For example: if rawfolder = example/project, and the years are in example/project/year, then it will work. A caution when calculating on a large server, this program initially gathers all file information before copying, therefore it may take a long time to initiate on a cluttered server
}

