% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exodus_fit_spline_comb.R
\name{exodus_fit_spline_comb}
\alias{exodus_fit_spline_comb}
\title{Fits exodus curves using the combined method for a series of files}
\usage{
exodus_fit_spline_comb(out_netcdf, radar, max_min = 15, per_rate = 0.15,
  nexrad_n = "V:/Documents/nexrad_site_list_with_utm.csv")
}
\arguments{
\item{out_netcdf}{location of reflectivity file created from exodusangle_netcdf}

\item{radar}{3 letter radar code}

\item{max_min}{maximum minutes to consider for curve fitting}

\item{per_rate}{percentage of change in reflectivity to consider the start of exodus}

\item{nexrad_n}{location of the nexrad site table. Default is the location on the AP's server}
}
\description{
Fits exodus curves using the combined method for a series of files
}

