% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{extractInput,MeRIP.RADAR-method}
\alias{extractInput,MeRIP.RADAR-method}
\title{extractINPUT}
\usage{
\S4method{extractInput}{MeRIP.RADAR}(object, normalized = TRUE)
}
\arguments{
\item{object}{The MeRIP.RADAR object}

\item{normalized}{logical option, whether to return normalized IP read counts. Default is TRUE}
}
\description{
The extractor to return the INPUT read count matrix of the experiment.
}
