% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{extractIP,MeRIP.RADAR-method}
\alias{extractIP,MeRIP.RADAR-method}
\title{extractIP}
\usage{
\S4method{extractIP}{MeRIP.RADAR}(object, normalized = TRUE,
  adjusted = FALSE, filtered = FALSE)
}
\arguments{
\item{object}{The MeRIP.RADAR object}

\item{normalized}{logical option, whether to return normalized IP read counts. Default is TRUE.}

\item{adjusted}{logical option, whether to return normalized and expression level adjusted IP read counts. Default is FALSE.}

\item{filtered}{logical option, whether to return normalized, adjusted and low-read-count-filtered count matrix. Default is FALSE.}
}
\description{
The extractor to return the IP read count matrix of the experiment.
}
