% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{features_to_dictionary}
\alias{features_to_dictionary}
\title{Retrieve the description of a list of input features}
\usage{
features_to_dictionary(rdr = NULL, feature_names = NULL)
}
\arguments{
\item{rdr}{A RadAR object (class \code{\link{SummarizedExperiments}})}

\item{feature_names}{(character) names of input features.}
}
\value{
A table of class \code{\link{tibble}} reporting feature names and dictionary.
}
\description{
Retrieve the description of a list of input features
}
\examples{

}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
