% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_functions.R
\name{filter_by_feature_type}
\alias{filter_by_feature_type}
\title{Filter a RadAR object by feature type(s)}
\usage{
filter_by_feature_type(rdr = NULL, feature_type = NULL)
}
\arguments{
\item{rdr}{A RadAR object (class \code{\link{SummarizedExperiment}})}

\item{image_type}{(character) Only features corresponding to feature_type will be retained.
Available values for feature_type depend on the software used for feature extraction.
More than one feature_type can be passed. Use \code{print_feature_type} to visualize available feature_type.}
}
\value{
An updated rdr (a RadAR object)
}
\description{
Filter a RadAR object by feature type(s)
}
\examples{

}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
