% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_functions.R
\name{filter_by_image_type}
\alias{filter_by_image_type}
\title{Filter a RadAR object by image type(s)}
\usage{
filter_by_image_type(rdr = NULL, image_type = NULL)
}
\arguments{
\item{rdr}{A RadAR object (class \code{\link{SummarizedExperiment}})}

\item{image_type}{(character) Only features corresponding to image_type will be retained.
Available values for image_type depend on the software used for feature extraction.
More than one image_type can be passed. Use \code{print_image_type} to visualize available image_type.}
}
\value{
An updated rdr (a RadAR object)
}
\description{
Filter a RadAR object by image type(s)
}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
