% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{filterBins,MeRIP.RADAR-method}
\alias{filterBins,MeRIP.RADAR-method}
\title{filterBins}
\usage{
\S4method{filterBins}{MeRIP.RADAR}(object, minCountsCutOff = 10)
}
\arguments{
\item{minCountsCutOff}{The minimal read count required in each bin, default is 10. This cutoff}

\item{x}{The MeRIP.RADAR object}
}
\description{
filterBins
}
