% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterBins.R
\name{filterBins}
\alias{filterBins}
\title{filterBins}
\usage{
filterBins(x, minCountsCutOff = 10)
}
\arguments{
\item{x}{The RNADMethyl data list}

\item{minCountsCutOff}{The minimal read count required in each bin, default is 10. This cutoff}
}
