% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_functions.R
\name{find_clusters}
\alias{find_clusters}
\title{Flag samples and/or features by cluster membership}
\usage{
find_clusters(rdr = NULL, n_clusters_cols = NULL,
  n_clusters_features = NULL)
}
\arguments{
\item{rdr}{A RadAR object (class \code{\link{SummarizedExperiment}})}

\item{n_clusters_cols}{(numeric) Number of clusters for patients/ROIs (columns)}

\item{n_clusters_features}{(numeric) Number of clusters for features (rows)}
}
\value{
An updated rdr (a RadAR object)
}
\description{
This function assigns cluster membership to patients/ROIs and/or features
based on previous hierarchical clustering (\code{\link{do_hierarchical_clustering}})
}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
