% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_spline_comb.R
\name{fit_spline_comb}
\alias{fit_spline_comb}
\title{Fits a single exodus curve using the combined method}
\usage{
fit_spline_comb(start_t, data, nspline = round((max(data$d_time) -
  min(data$d_time)) * 30), method = "natural", per_rate = 0.2,
  max_min = 15)
}
\arguments{
\item{start_t}{Approximate start time}

\item{data}{A data.frame with two columns: d_time = decimal time and ref = reflectivity.}

\item{nspline}{Number of spline points, defaults to 2 minute intervals}

\item{method}{The method for spline fitting. This doesn't change fitting much}

\item{per_rate}{Percentage of change in reflectivity to consider the start of exodus}

\item{max_min}{Maximum minutes to consider for curve fitting}
}
\description{
Prints out the curve, sunset, and what time was fit based on user specific criteria
}

