% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{geneExpression}
\alias{geneExpression}
\title{extractor for RNAseq data}
\usage{
\S4method{geneExpression}{MeRIP.RADAR}(object)
}
\arguments{
\item{object}{The MeRIP object}
}
\description{
Extract gene level expression (RNAseq) data in normalized read counts
}
