% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_movement.R
\name{generic_movement}
\alias{generic_movement}
\title{Append generic movement parameters}
\usage{
generic_movement(local)
}
\arguments{
\item{local}{the connection to the local database}
}
\description{
Append generic movement parameters
}
