% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{geom}
\alias{geom}
\title{Calculates the geometric mean}
\usage{
geom(x, na.rm = T)
}
\arguments{
\item{x}{vector of values}

\item{na.rm}{removes NA values automatically, Default=T}
}
\description{
Calculates the geometric mean
}

