\name{GeometryHalfPowerRadius}
\alias{GeometryHalfPowerRadius}
\title{Half-power radius}
\usage{GeometryHalfPowerRadius(r, bwhalf)}
\description{\code{GeometryHalfPowerRadius} returns the half-power radius
Battan (1973)
}
\arguments{
  \item{r}{Range [m]}
  \item{bwhalf}{Half-power beam width [degrees]}
}
\value{
  \item{Rhalf}{ Half-power radius [m]}
}
\source{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}

Louis J. Battan, 1973
Radar Observation of the Atmosphere
University of Chicago Press
}
\references{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}

Louis J. Battan, 1973
Radar Observation of the Atmosphere
University of Chicago Press
}
\author{Jose Gama}
\keyword{programming}
