% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracks_around_centre.R
\name{get_tracks_around_centre}
\alias{get_tracks_around_centre}
\title{Extract tracks nearest to a given position}
\usage{
get_tracks_around_centre(
  local,
  centre,
  max_distance = 200,
  max_height = 200,
  max_tracks = 100,
  relevant = 1,
  rate = 2
)
}
\arguments{
\item{local}{the connection to the local database}

\item{centre}{a data frame with a single row and x and y coordinates of the
centre.
Coordinates in CRS 31370}

\item{max_distance}{Use all equal time tracks where the minimum distance in
meters between the track and the centre is smaller than \code{max_distance}.}

\item{max_height}{Only use track points with an altitude lower than
\code{max_height}.
Defaults to 200 meter.}

\item{max_tracks}{Return only the nearest \code{max_tracks} tracks.}

\item{relevant}{The value of the \code{relevant} variable of the \code{species} table
in the database.
Select only tracks from species with a value equal or larger than this
number.}

\item{rate}{The time difference between points to calculate the step length
and turning angles.}
}
\description{
Extract tracks nearest to a given position
}
