% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_functions.R
\name{import_3dslicer}
\alias{import_3dslicer}
\title{Import 3DSlicer data}
\usage{
import_3dslicer(dir)
}
\arguments{
\item{dir}{(character) Path to the directory containing tsv files generated by 3DSlicer}
}
\value{
A RadAR object (class \code{\link{SummarizedExperiment}})
}
\description{
This functions imports feature values extracted using 3DSlicer software
}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
