% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_functions.R
\name{import_lifex}
\alias{import_lifex}
\title{Import LifeX data}
\usage{
import_lifex(dir)
}
\arguments{
\item{dir}{(character) Path to the directory containing xls or csv files generated by LifeX}
}
\value{
A RadAR object (class \code{\link{SummarizedExperiment}})
}
\description{
Import LifeX data
}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
