% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_functions.R
\name{import_lifex_session}
\alias{import_lifex_session}
\title{Import LifeX session data}
\usage{
import_lifex_session(session)
}
\arguments{
\item{session}{(character) Path to the session file generated by LifeX}
}
\value{

}
\description{
Import LifeX session data
}
