% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_functions.R
\name{import_pyradiomics}
\alias{import_pyradiomics}
\title{Import pyradiomics data}
\usage{
import_pyradiomics(dir = NULL)
}
\arguments{
\item{dir}{(character) Path to the directory containing tsv files generated by pyradiomics}
}
\value{
A RadAR object (class \code{\link{SummarizedExperiment}})
}
\description{
This functions imports feature values extracted using pyradiomics software
}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
