% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_functions.R
\name{import_radiomic_table}
\alias{import_radiomic_table}
\title{Import features}
\usage{
import_radiomic_table(file = NULL)
}
\arguments{
\item{file}{(character) Path to the table of radiomic features. It should be tab delimited.
First row should report sample names. First column should report feature names.}
}
\value{
A RadAR object (class \code{\link{SummarizedExperiment}})
}
\description{
This functions imports a table of radiomic features (features in rows, samples in columns)
}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
