% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{incomp}
\alias{incomp}
\title{Incomparables: A function that compares one value against a vector of options and returns the closest value to it.}
\usage{
incomp(x, comp)
}
\arguments{
\item{x}{vector of values to match against}

\item{comp}{the singular number to match}
}
\description{
The function is primarily for situations where the numbers are not exact matches. Does not distinguish between positive or negative differences.
}

