\name{kConstantSpeedOfLight}
\alias{kConstantSpeedOfLight}
\alias{kConstantSLP}
\alias{kConstantP0}
\alias{kConstantRe}
\alias{kConstantBoltz}
\alias{kConstantR43}
\title{Constant speed of light }
\usage{kConstantSpeedOfLight}
\description{\code{kConstantSpeedOfLight} is "c" the constant speed of light [m/s].

\code{kConstantSLP} Sea-level Pressure [hPa].

\code{kConstantP0} Reference pressure [hPa].

\code{kConstantRe} Earth's radius [m].

\code{kConstantR43} 4/3 Approximation effective radius for standard atmosphere [m].

\code{kConstantBoltz} Boltzmann's constant [ m^2 kg s^-2 K^-1].
}
\author{Jose Gama}
\examples{
print(kConstantSpeedOfLight)
}
\keyword{datasets}
