% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_database.R
\name{local_database}
\alias{local_database}
\title{Connect to a local SQLite database
Connects to the \code{radar.sqlite} database}
\usage{
local_database(path = ".")
}
\arguments{
\item{path}{the path of the database}
}
\description{
Connect to a local SQLite database
Connects to the \code{radar.sqlite} database
}
