% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vpr.R
\name{make_vpr}
\alias{make_vpr}
\title{Makes VPR's from interpolated reflectivity files}
\usage{
make_vpr(radar, basegrid_n, vprdir, refdir, radio_fold, radiosonde, ...,
  nexrad_n = "V:/Documents/nexrad_site_list_with_utm.csv",
  mask = elevframe$id, z_min = 0, z_max = 3165, start = 1,
  end = length(files), minvprrange = 5000, maxvprrange = 20000,
  vprbottom = 10, vprtop = 1500, overwrite = T, sun = F,
  elev_min = 0.5, elev_max = 5)
}
\arguments{
\item{radar}{The 3 letter radar code}

\item{basegrid_n}{The location of the basegrid}

\item{vprdir}{Location of the vpr files}

\item{refdir}{Location of the reflectivity files}

\item{radio_fold}{Location of the radiosondes, if applicable}

\item{nexrad_n}{Location of the nexrad site table. Default is the location on the AP's server}

\item{mask}{A vector of the ID of pulse volumes you want considered in the VPR calculation}

\item{z_min}{The minimum reflectivity considered for vpr calculation. Filtered out based on the 0.5 angle.}

\item{z_max}{The maximum reflectivity considered for vpr calculation.}

\item{start}{An override value, describing what number of interpolated reflectivity files to start from}

\item{end}{An override value, described what number of interpolated reflectivity files to end from}

\item{minvprrange}{The minimum range from the radar to consider for vpr calculation (in meters)}

\item{maxvprrange}{The maximum range from the radar to consider for vpr calculation (in meters)}

\item{vprbottom}{The lowest height bin to calculate vpr values}

\item{vprtop}{The heighst height bin to calculate vpr values}

\item{overwrite}{Whether you'd like to overwrite already made VPRS or skip them}

\item{sun}{Filter out approximate location of sunflares in vpr calculation, only does so for angles 87 - 93}

\item{elev_min}{The lowest beam elevation to consider for vpr calculation}

\item{elev_max}{The heighest beam elevation to consider. It is safer to put a value higher than the heighest to be considered}

\item{radisonde}{Does the radar have radiosonde}
}
\description{
The majority of these arguments are special case scenarios and have default values. The only ones required are: radar, basegridn_n, vprdir, refdir, radio_fold, and radiosonde
}

