% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{mround}
\alias{mround}
\title{Rounds to any decimal or whole value}
\usage{
mround(x, dec)
}
\arguments{
\item{x}{vector of values}

\item{dec}{the number/decimal you want to convert to.}
}
\description{
Rounds to any decimal or whole value
}

