% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NARR_get.R
\name{NARR_get}
\alias{NARR_get}
\title{Downloads lateast NARR wind data (North American Regional Reanalysis)}
\usage{
NARR_get(radar, narr_dir, date_time, ...,
  nexradmatch = "V:/Documents/RadarData/nexradmatch.dbf",
  xy_n = "V:/Documents/RadarData/xy.csv",
  alternatesource = "V:/Documents/RadarData/alternateNARR")
}
\arguments{
\item{radar}{The 3 letter radar code}

\item{narr_dir}{The output directory for the summarized narr files}

\item{date_time}{The date and time of files you want downloaded, in the correct 3 hour segment and in as.POSIX date class}

\item{nexradmatch}{Location of a file of NARR cells that match with specified radar sites. Default location is on the AP's server}

\item{xy_n}{Location of a file of x and y coordinates of the narr files. Default location is on the AP's server}

\item{alternatesource}{Location of files later than December 2014 that have to be manually downloaded. Default location is on the AP's server.}
}
\description{
Currently only downloads wind data for the 1000mb - 600mb layer. If you'd like to download other variables, NCEP data from the RNCEP package should be adequate. 
This program only works when supplying specific radars which are detailed in the nexradmatch file. We don't currently have a program made that allows you to download based on lat/longs like the RNCEP package.
}

