% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NARR_vpr.R
\name{NARR_vpr}
\alias{NARR_vpr}
\title{Downloads lateast NARR files (North American Regional Reanalysis) using the vpr files as a template.}
\usage{
NARR_vpr(radar, narr_dir, vprdir, dates = vpr_dates, ...,
  nexradmatch = "V:/Documents/RadarData/nexradmatch.dbf",
  xy_n = "V:/Documents/RadarData/xy.csv",
  alternatesource = "V:/Documents/RadarData/alternateNARR")
}
\arguments{
\item{radar}{The 3 letter radar code}

\item{narr_dir}{The output directory for the summarized narr files}

\item{vprdir}{The file path of the vpr files (searches for the pattern: "VprTable_")}

\item{dates}{Override argument. A vector of the dates to consider in the vprfolder. In the format: yyyy-mm-dd}

\item{nexradmatch}{Location of a file of NARR cells that match with specified radar sites. Default location is on the AP's server}

\item{xy_n}{Location of a file of x and y coordinates of the narr files. Default location is on the AP's server}

\item{alternatesource}{Location of files for dates after December 2014 that have to be manually downloaded. Default location is on the AP's server.}
}
\description{
Times are bracketed around the vprs time.
}

