% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexrad_get.R
\name{nexrad_get}
\alias{nexrad_get}
\title{Automatically grabs variables related to specific radar calculations}
\usage{
nexrad_get(radar, nexrad_n = "V:/Documents/nexrad_site_list_with_utm.csv")
}
\arguments{
\item{radar}{The 3 letter radar code}

\item{nexrad_n}{Location of the nexrad site table. Default is the location on the AP's server}
}
\description{
This program will create variables in the global environment that are typically used for radar calculations. They include latitude, longitude, tower height, antenna height, and earth radius from the radar tower.
}

