% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_functions.R
\name{normalize_feature_values}
\alias{normalize_feature_values}
\title{Apply normalization to feature values}
\usage{
normalize_feature_values(
  rdr = NULL,
  method = "quantile",
  which_data = "scaled"
)
}
\arguments{
\item{rdr}{A RadAR object (class \code{\link{SummarizedExperiment}})}

\item{method}{Only quantile normalization has been implemented so far. After quantile normalization,
the distributions of feature values are identical in terms of statistical properties}

\item{which_data}{(character) Which data should be normalized. It can be one of "normal" or "scaled".}
}
\value{
An updated rdr (a RadAR object)
}
\description{
Feature values are quantile normalized by limma package.
After quantile normalization, the distributions of feature values are identical in terms of
statistical properties.
}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
