% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeLibrary.R
\name{normalizeLibrary}
\alias{normalizeLibrary}
\title{normalizeLibrary}
\usage{
normalizeLibrary(reads, samplenames, X)
}
\arguments{
\item{reads}{Read counts matrix from countReads() function.}

\item{samplenames}{A vector of Samples names. Needs to be the same as the one input to countReads()}

\item{X}{Study design or Grouping for the samples, should have 2 levels}
}
