% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_windmill.R
\name{osm_windmill}
\alias{osm_windmill}
\title{Get windmill locations from OpenStreetMap}
\usage{
osm_windmill(
  local,
  bbox = c(xmin = 4.26, ymin = 51.22, xmax = 4.5, ymax = 51.35)
)
}
\arguments{
\item{local}{the connection to the local database}

\item{bbox}{Either (i) four numeric values specifying the maximal and minimal
longitudes and latitudes, in the form \code{c(xmin, ymin, xmax, ymax)}
or (ii) a character string in the form \code{xmin,ymin,xmax,ymax}. These
will be passed to \link[osmdata]{getbb} to be converted to a numerical bounding
box. Can also be (iii) a matrix representing a bounding polygon as
returned from \code{getbb(..., format_out = "polygon")}. To search in an
area, (iv) a character string with a relation or a (closed) way id in
the format \code{"way(id:1)"}, \code{"relation(id:1, 2)"} or \code{"relation(id:1, 2, 3); way(id:2)"} as returned by \code{getbb(..., format_out = "osm_type_id")}
or \link[osmdata]{bbox_to_string} with a \code{data.frame} from \code{getbb(..., format_out = "data.frame")} to select all areas combined (relations and ways).}
}
\description{
Download the windmill information in the bounding box.
Store the location, height and operator in the local database.
}
