% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{peakDistribution}
\alias{peakDistribution}
\title{plot distribution of peaks on gene annotation}
\usage{
\S4method{peakDistribution}{MeRIP.RADAR}(object)
}
\arguments{
\item{object}{The MeRIP.RADAR object}
}
\description{
plot distribution of differential peaks on gene annotation
}
