% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_correlation_matrix}
\alias{plot_correlation_matrix}
\title{Draw correlation matrix of radiomic features}
\usage{
plot_correlation_matrix(rdr = NULL, view_as = c("heatmap", "corrplot"),
  method_correlation = "pearson", method_hcl = "ward.D",
  which_data = "normal", cex_labels = 0.5, srt_labels = 45,
  ncolors = 50, save_plot_to = NULL, width_plot = 20,
  height_plot = 15, show_names = T, show_dendrogram = T)
}
\arguments{
\item{rdr}{A RadAR object (class \code{\link{SummarizedExperiment}}).}

\item{view_as}{(character) How to visualize correlation matrix? It can be "corrplot" or "heatmap".}

\item{method_correlation}{(character) Which method use to calculate correlation. It can be one of the following:
"pearson", "kendall", "spearman".}

\item{method_hcl}{(character) Which method use for clustering matrix columns and rows. Available methods are
"ward.D", "ward.D2", "single", "complete","average" , "mcquitty","median", "centroid".}

\item{which_data}{(character) Which data use for the computation of the correlation coefficients. It can be one of the following: "normal",
"scaled", "normalized".}

\item{cex_labels}{(numeric) Text label cex.}

\item{srt_labels}{(numeric) Text label string rotation in degrees.}

\item{ncolors}{(numeric) Number of colors to generate the palette.}

\item{save_plot_to}{(character) If specified, filename where the heatmap will be saved to.}

\item{width_plot}{(numeric) output file width in inches.}

\item{height_plot}{(numeric) output file height in inches.}

\item{show_names}{(logical) Should feature names (labels) be visualized?}

\item{show_dendrogram}{(logical) Should dendogram (for columns and rows) be drawn (valid for heatmap visualzation)?}
}
\value{
none
}
\description{
This function draws a correlation matrix of the radiomic dataset by \code{\link{corrplot}} or \code{\link{pheatmap}}.
Correlation matrix can be visualized by clustered correlation plot or heatmap.
}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
