% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_features}
\alias{plot_features}
\title{Draw boxplot + stripchart of selected feature(s)}
\usage{
plot_features(
  rdr = NULL,
  feature_names = NULL,
  conditions = NULL,
  which_data = "scaled",
  max_range = F,
  user_range = NULL
)
}
\arguments{
\item{rdr}{A RadAR object (class \code{\link{SummarizedExperiment}}).}

\item{feature_names}{(character) Which feature(s) should be plotted. Available feature names are
obtained by \code{rowData(rdr)$feature_name}.}

\item{conditions}{(numeric, character) Vector of labels specifying conditions for each sample in rdr.
Should have the same length of \code{ncol(rdr)}.}

\item{which_data}{(character) Which data use for plot. It can be one of the following: "normal", "scaled", "normalized".}

\item{max_range}{(logical) Should plot(s) be ranged between min and max values?}

\item{user_range}{(numeric) A vector of length 2 reporting min and max ylim values.}
}
\value{
none
}
\description{
This function generates boxplots + stripcharts of selected feature(s). Data are stratified by
user defined conditions.
}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
