% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_heatmap_hcl}
\alias{plot_heatmap_hcl}
\title{Draw a clustered heatmap of the radiomic dataset}
\usage{
plot_heatmap_hcl(
  rdr = NULL,
  show_feature_names = F,
  show_sample_names = F,
  annotation_tracks = NULL,
  annotation_data = NULL,
  annotation_colors = NULL,
  ncolors = 50,
  save_plot_to = NULL,
  width_plot = 20,
  height_plot = 15
)
}
\arguments{
\item{rdr}{A RadAR object (class \code{\link{SummarizedExperiment}}).}

\item{show_feature_names}{(logical) Should feature names be displayed in the heatmap?}

\item{show_sample_names}{(logical) Should sample names be displayed in the heatmap?}

\item{annotation_tracks}{(character) Tracks to report in the annotation section of the heatmap.
Available annotation tracks can be obtained by \code{\link{names(colData(rdr))}}.}

\item{annotation_data}{(data.frame) A data frame specifying tracks to report in the annotation section. For
more details refer to \code{\link{?pheatmap::pheatmap}}.}

\item{annotation_colors}{(list) List for specifying annotation_col track colors manually.For
more details and examples refer to \code{\link{?pheatmap::pheatmap}}.}

\item{ncolors}{(integer) Number of colors of the color palette.}

\item{save_plot_to}{(character) If specified, filename where the heatmap will be saved to.}

\item{width_plot}{(numeric) output file width in inches.}

\item{height_plot}{(numeric) output file height in inches.}
}
\value{
none
}
\description{
This function uses \code{\link{pheatmap}} to draw a heatmap of feature values (rows) across samples (columns). The rows and columns of the
heatmap are ordered based on previous required hierarchical clustering generated by \code{\link{do_hierarchical_clustering}}.
Dendrograms for columns and rows are also reported with optional annotation tracks.
}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
