% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneRADAR.R
\name{plotGene}
\alias{plotGene}
\title{plotGene}
\usage{
plotGene(IP_BAM, INPUT_BAM, size.IP, size.INPUT, geneName, geneModel,
  libraryType = "opposite", center = mean, GTF, ZoomIn = NULL)
}
\arguments{
\item{IP_BAM}{The bam files for IP samples}

\item{INPUT_BAM}{The bam files for INPUT samples}

\item{size.IP}{The size factor for IP libraries}

\item{size.INPUT}{The size factor for INPUT libraries}

\item{geneName}{The name (as defined in gtf file) of the gene you want to plot}

\item{geneModel}{The gene model generated by gtfToGeneModel() function}

\item{libraryType}{"opposite" for mRNA stranded library, "same" for samll RNA library}

\item{GTF}{gtf annotation as GRanges object. Can be obtained by GTF <- rtracklayer::import("xxx.gtf",format = "gtf")}
}
