% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{plotGeneCov,MeRIP.RADAR-method}
\alias{plotGeneCov,MeRIP.RADAR-method}
\title{plotGeneCov}
\usage{
\S4method{plotGeneCov}{MeRIP.RADAR}(object, geneName,
  libraryType = "opposite", center = mean, ZoomIn = NULL,
  adjustExprLevel = FALSE, split = FALSE)
}
\arguments{
\item{object}{The MeRIP. object}

\item{geneName}{The gene symbol to be ploted.}

\item{libraryType}{Specify whether the library is the same or opposite strand of the original RNA molecule. Default is "opposite".}

\item{center}{Specify the method to calculate average coverage of each group. Could be mean or median.}

\item{ZoomIn}{c(start,end) The coordinate to zoom in at the gene to be ploted.}

\item{adjustExprLevel}{logical parameter. Specify whether normalize the two group so that they have similar expression level.}

\item{split}{Logical option. Whether to split plots for each individual (TRUE), or plot each group by group mean/median coverage (FALSE, default).}

\item{GTF}{The GRanges object containing gtf annotation. Can obtain by rtracklayer::import("file.gtf", format= "gtf").}
}
\description{
plotGeneCov
}
