% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneCov.R
\name{plotGeneCoverage}
\alias{plotGeneCoverage}
\title{plotGeneCoverage}
\usage{
plotGeneCoverage(IP_BAMs, INPUT_BAMs, size.IP, size.INPUT, X, geneName,
  geneModel, libraryType = "opposite", center = mean, GTF, ZoomIn = NULL,
  adjustExprLevel = FALSE, plotSNP = NULL)
}
\arguments{
\item{size.IP}{The size factor for IP libraries}

\item{size.INPUT}{The size factor for INPUT libraries}

\item{geneName}{The name (as defined in gtf file) of the gene you want to plot}

\item{geneModel}{The gene model generated by gtfToGeneModel() function}

\item{libraryType}{"opposite" for mRNA stranded library, "same" for samll RNA library}

\item{GTF}{gtf annotation as GRanges object. Can be obtained by GTF <- rtracklayer::import("xxx.gtf",format = "gtf")}

\item{adjustExprLevel}{Logic parameter determining whether adjust coverage so that input are at "same" expression level.}

\item{plotSNP}{The option to plot SNP on the figure. Null by default. If want to include SNP in the plot, the parameter needs to ba a dataframe like this:  data.frame(loc= position, anno="A/G")}

\item{IP_BAM}{The bam files for IP samples}

\item{INPUT_BAM}{The bam files for INPUT samples}
}
