% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneRADAR.R
\name{plotGenePair}
\alias{plotGenePair}
\title{plotGenePair}
\usage{
plotGenePair(Ctl_IP_BAM, Ctl_INPUT_BAM, Treat_IP_BAM, Treat_INPUT_BAM,
  Ctl_size.IP, Ctl_size.INPUT, Treat_size.IP, Treat_size.INPUT, geneName,
  geneModel, libraryType = "ooposite", center = mean, GTF, ZoomIn = NULL)
}
\arguments{
\item{Ctl_IP_BAM}{The bam files for Control IP samples}

\item{Ctl_INPUT_BAM}{The bam files for Control INPUT samples}

\item{Treat_IP_BAM}{The bam files for treated IP samples}

\item{Treat_INPUT_BAM}{The bam files for treated INPUT samples}

\item{Ctl_size.IP}{The size factor for IP libraries}

\item{Ctl_size.INPUT}{The size factor for INPUT libraries}

\item{Treat_size.IP}{The size factor for IP libraries}

\item{Treat_size.INPUT}{The size factor for INPUT libraries}

\item{geneName}{The name (as defined in gtf file) of the gene you want to plot}

\item{geneModel}{The gene model generated by gtfToGeneModel() function}

\item{libraryType}{"opposite" for mRNA stranded library, "same" for samll RNA library}
}
\description{
plot tow groups of samples in the same figure
}
