% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneRADAR.R
\name{plotGeneRADAR}
\alias{plotGeneRADAR}
\title{plotGeneMonster}
\usage{
plotGeneRADAR(readsOut, geneName, libraryType = "opposite", center = "mean",
  GTF, ZoomIn = NULL)
}
\arguments{
\item{readsOut}{The data list from countReads and other analysis.}

\item{geneName}{The gene symbol to be ploted.}

\item{GTF}{The GRanges object containing gtf annotation.}

\item{ZoomIn}{c(start,end) The coordinate to zoom in at the gene to be ploted.}
}
