% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{plotHeatMap}
\alias{plotHeatMap}
\title{plotHeatMap}
\usage{
\S4method{plotHeatMap}{MeRIP.RADAR}(object, covariates = TRUE)
}
\arguments{
\item{object}{The MeRIP.RADAR object.}

\item{covariates}{Logic parameter. Whether to regress out covariates (if variable has been set to have more than one column). Default is TRUE.}
}
\description{
Plot heat map of methylation variations across samples been analyzed.
}
