% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plotTPM}
\alias{plotTPM}
\title{plotTPM}
\usage{
plotTPM(TPM, geneName, group, logCount = FALSE, facet_grid = FALSE)
}
\arguments{
\item{TPM}{Dataframe of gene TPM}

\item{geneName}{The name of genes to be ploted.}

\item{group}{Categorical info for each sample.}

\item{logCount}{where to plot count at log scale}
}
\description{
plotTPM
}
