% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{print_distance_methods}
\alias{print_distance_methods}
\title{Print available methods for computing distance}
\usage{
print_distance_methods()
}
\value{
A screen message with available methods.
}
\description{
Print available methods for computing distance
}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
