% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{print_feature_type}
\alias{print_feature_type}
\title{Print available feature types}
\usage{
print_feature_type(rdr = NULL)
}
\arguments{
\item{rdr}{A RadAR object (class \code{\link{SummarizedExperiments}})}
}
\value{
A screen message showing available feature types.
}
\description{
Print available feature types
}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
