% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{print_hcl_methods}
\alias{print_hcl_methods}
\title{Print available methods for hierarchical clustering}
\usage{
print_hcl_methods()
}
\value{
A screen message with available methods.
}
\description{
Print available methods for hierarchical clustering
}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
