% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{print_image_type}
\alias{print_image_type}
\title{Print available image types}
\usage{
print_image_type(rdr = NULL)
}
\arguments{
\item{rdr}{A RadAR object (class \code{\link{SummarizedExperiments}})}
}
\value{
A screen message showing available image types.
}
\description{
Print available image types
}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
