% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radio_get.R
\name{radio_get}
\alias{radio_get}
\title{Downloads radiosondes from the university of Wyoming website for a radar in .html format}
\usage{
radio_get(dates, output, radar, nexrad_n, years = unique(nights2$year) + 1900)
}
\arguments{
\item{dates}{The Dates to be downloaded. In the format mm/dd/yyyy. These are converted to the date at midnight of the next day}

\item{output}{Location to save files}

\item{radar}{The 3 letter radar code}

\item{nexrad_n}{Location of the nexrad site table. Default is the location on the AP's server}

\item{years}{The years you want to download. Default is all years in the dates list.}
}
\description{
This program uses site ID's taken from the University of Wyoming that are stored in the nexradsite table. You may need to update that file to download radiosondes. Output is in .html.
}

