% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_azim.R
\name{re_azim}
\alias{re_azim}
\title{Resamples azimuths}
\usage{
re_azim(data, old_azim, new_azim = seq(0, 359.5, 0.5))
}
\arguments{
\item{data}{A dataframe needing to be resampled}

\item{old_azim}{Vector of old_azimuths that matches the data}

\item{new_azim}{A sequence of azimuths to resample data to}
}
\description{
Resamples azimuths
}

