% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recal_prog.R
\name{recal_corr}
\alias{recal_corr}
\title{Calculates the correction factor for a specific range, groundht, and pblock.}
\usage{
recal_corr(data, vpr, elev)
}
\arguments{
\item{data}{A dataframe of all necessary information needed for calculation. Includes 6 columns: tbeam = top of the beam, mbeam = middle of the beam, bbeam = bottom of the beam, pblock = percent of beam blocked, range = range from radar in meters, groundht = ground height above sealevel.}

\item{vpr}{A dataframe of the VPR. Includes 2 columns: hght = height in meters, zh1 = the calculated vpr value}
}
\description{
Used primarily in the summar code and may not be very useful as a stand alone program. Calculated using standard refraction.
}

