% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPoissonGammaMerge.R
\name{reportPoissonGammaMerge}
\alias{reportPoissonGammaMerge}
\title{reportMergeBin}
\usage{
reportPoissonGammaMerge(stats, geneGRList, cutoff, binSize = 50)
}
\arguments{
\item{stats}{read count mattrix}

\item{geneGRList}{GRange List of gene model from gtfToGeneModel(). Or gtf file.}

\item{cutoff}{The p_value cutoff to merge the bin}
}
