% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{reportResult,MeRIP.RADAR-method}
\alias{reportResult,MeRIP.RADAR-method}
\title{reportResult}
\usage{
\S4method{reportResult}{MeRIP.RADAR}(object, cutoff = 0.1,
  Beta_cutoff = 0.5, threads = 1)
}
\description{
merge significant bins and report result as BED12 format
}
