% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{results}
\alias{results}
\title{export results}
\usage{
\S4method{results}{MeRIP.RADAR}(object)
}
\arguments{
\item{object}{The MeRIP.RADAR object.}
}
\value{
joint peaks (with test result) in a data.frame.
}
\description{
The extractor for final test result.
}
