% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_functions.R
\name{scale_feature_values}
\alias{scale_feature_values}
\title{This function implements different scaling strategies to feature values}
\usage{
scale_feature_values(rdr = NULL, method = "minmax")
}
\arguments{
\item{rdr}{A RadAR object (class \code{\link{SummarizedExperiment}})}

\item{method}{(character) Which method use for scaling feature values. It can be one of "minmax" or "median".
Using min-max, all features will take values in the range [0,1].
Using median, all features will be median subctracted.}
}
\value{
An updated rdr (a RadAR object)
}
\description{
This function implements different scaling strategies to feature values
}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
