% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{select}
\alias{select}
\title{subset MeRIPdata}
\usage{
\S4method{select}{MeRIP}(object, samples)

\S4method{select}{MeRIP.RADAR}(object, samples)
}
\arguments{
\item{object}{The MeRIP object}

\item{samples}{The samplenames to be subset or the index number of samples to be subset.}

\item{object}{The MeRIP.RADAR object}

\item{samples}{The samplenames to be subset or the index number of samples to be subset.}
}
\value{
an MeRIP object of selected samples.

an MeRIP.RADAR object of selected samples.
}
\description{
subset dataset by samples.

subset dataset by samples.
}
