% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{select_top_features}
\alias{select_top_features}
\title{Select top radiomic features according to a given statistics}
\usage{
select_top_features(
  rdr = NULL,
  which_statistics = "wilcox",
  thr_pvalue = 0.05,
  thr_auc = 0.8,
  thr_concordance = 0.1,
  thr_cox_zvalue = 3,
  write_to = NULL
)
}
\arguments{
\item{rdr}{A RadAR object (class \code{\link{SummarizedExperiment}}).}

\item{which_statistics}{(character) Select top features based on one of the following pre-computed statistics:
"wilcox", "AUC", "concordance" or "cox".}

\item{thr_pvalue}{(numeric) P-value threshold to identify statistically significant features from wilcox.
It should be in the range (0, 1].}

\item{thr_auc}{(numeric) AUC threshold to identify statistically significant features from AUC
It should be in the range (0.5, 1].}

\item{thr_concordance}{(numeric) Threshold to identify statistically significant features from concordance analysis.
It should be in the range (0, .5].}

\item{thr_cox_zvalue}{(numeric) Z threshold (Wald statistics) to identify statistically significant features from cox regression analysis.
It should be in the range (0, inf).}

\item{write_to}{(character) If specified, filename to output top statistically significant features (tab-delimited).}
}
\value{
A table of class \code{\link{tibble}} reporting top features.
}
\description{
This function create a summary table of most significant features based on previously computed statistics
(wilcox, AUC, concordance index or cox regression)
}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
