% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{sem}
\alias{sem}
\title{Calculates the standard error of the mean}
\usage{
sem(x, na.rm = T)
}
\arguments{
\item{x}{vector of values}

\item{na.rm}{removes NA values automatically. Default=T}
}
\description{
Calculates the standard error of the mean
}

