% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_print.R
\name{spline_print}
\alias{spline_print}
\title{Graphs spline fitted exodus curves, with sunset included}
\usage{
spline_print(my.data)
}
\arguments{
\item{my.data}{a data.frame with two columns: d_time = decimal time and ref = reflectivity.}
}
\description{
This function is specific to the curve-fitting process and likely not of important use outside of this program
}

