% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_exodus}
\alias{summary_exodus}
\title{Creates a summary file for multiple Interpolated Reflectivity files that filters out detection distance}
\usage{
summary_exodus(radar, folder, basegrid_n, out, refdir, vprdir, zoption = 3,
  header = T, legacy = F, vprbottom = 10, vprtop = 1500,
  nexrad_n = "V:/Documents/nexrad_site_list_with_utm.csv", corrected = T,
  start = 1, end = length(files), sun = F, sum_angle = 0.5,
  col_time = F)
}
\arguments{
\item{radar}{The 3 letter radar code}

\item{folder}{The project name that will be tacked onto the name of the files}

\item{basegrid_n}{The location of the basegrid}

\item{out}{Folder to write sumamry file}

\item{refdir}{Location of the InterpolatedReflectivity files}

\item{zoption}{What units would you like to convert reflectivity too (1: Z, 2: cm2/km3, 3: cm2/hectare)}

\item{header}{Do the vprs contain a header?}

\item{legacy}{Is the input from legacy data?}

\item{vprbottom}{What is the height of the bottom of the vpr?}

\item{vprtop}{What is the height of the vpr?}

\item{corrected}{Are the files corrected, if F it will correct using the vpr}

\item{start}{Override function indicating the file number you want to begin running}

\item{end}{Override function indicating the file number you want to end on}

\item{sun}{Would you like to filter out the location of a possible sun flare?}

\item{sum_angle}{What angle from the Interpolated Reflectivity files do you want to create the summary for?}

\item{col_time}{Do you want the column names to include the time (in decimal hour)}

\item{vpdir}{Location of the BirdVPRfiles}
}
\description{
This function will insert NAs for correction factors <0.5, correct data using the vpr, and insert all nights into one large summary file thats merged with the basegrid.
}

