\name{SystemAntEffArea}
\alias{SystemAntEffArea}
\title{Antenna effective area}
\usage{SystemAntEffArea(G, lam)}
\description{\code{SystemAntEffArea} returns the antenna effective area
From Rinehart (1997), Eqn 4.5
}
\arguments{
  \item{G}{Antenna Gain [dB]}
  \item{lam}{Radar wavelength [m]}
}
\value{
  \item{Ae}{ Antenna effective area [unitless]}
}
\source{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}

R. E. Rinehart, 1997
Radar for Meteorologists
Rinehart Publishing
}
\references{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}

R. E. Rinehart, 1997
Radar for Meteorologists
Rinehart Publishing
}
\author{Jose Gama}
\keyword{programming}
