\name{SystemFreq}
\alias{SystemFreq}
\title{Frequency given wavelength}
\usage{SystemFreq(lam, speedOfLight)}
\description{\code{SystemFreq} Converts from wavelength to frequency 
}
\arguments{
  \item{lam}{Wavelength [m]}
  \item{speedOfLight}{speed of light}
}
\value{
  \item{f}{ Frequency [Hz]}
}
\source{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}
}
\references{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}
}
\author{Jose Gama}
\keyword{programming}
