% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SZA_pn.R
\name{SZA_pn}
\alias{SZA_pn}
\title{Calculates a positive/negative sunangle depending on time of day}
\usage{
SZA_pn(timein = Sys.time(), Lat = 50.910335, Lon = 11.56874)
}
\arguments{
\item{timein}{as.POSIX format of time}

\item{Lat}{latitude}

\item{Lon}{longitude}
}
\description{
This program converts the standard SZA function in RAtmosphere to one that calculates positive and negative sunangles. Positive sunangles are ones from noon - midnight, and negative angles are ones between midnight - noon
}

