% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{test_radiomic_signature}
\alias{test_radiomic_signature}
\title{Test radiomic signature by Cox regression model}
\usage{
test_radiomic_signature(
  rdr = NULL,
  surv_obj = NULL,
  signature = NULL,
  signature_rdr = NULL,
  which_data = "scaled"
)
}
\arguments{
\item{rdr}{A RadAR object (class \code{\link{SummarizedExperiment}}).}

\item{surv_obj}{An object of class \code{\link{Surv}}.}

\item{signature}{(character) vector of radiomic features defining the signature.}

\item{signature_rdr}{RadAR object (class \code{\link{SummarizedExperiment}}) obtained by
\code{\link{do_feature_selection}}. Required to test a model generated using
 glmnet-cox or glmnet-binom.}

\item{which_data}{(character) Which data use to compute concordance index. It can be one of the following: "normal",
"scaled", "normalized".}
}
\value{
In case of mRMR, hcl and pca methods, the function returns the results of the cox regression model, including the concordance index.
In case of glmnet-cox method, the function returns a list of two elements, including 1) the prediction of the glmnet model on the new data
and 2) the computation of concordance index.
}
\description{
Test radiomic signature by Cox regression model
}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
