% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{time_diff}
\alias{time_diff}
\title{This gives you a unit of time in class 'time'}
\usage{
time_diff(number, unit)
}
\arguments{
\item{number}{a numeric of how many of time units you want}

\item{unit}{the unit of time accepts 'sec', 'min', 'hour', or 'day'}
}
\description{
This function is useful when you need to find the difference in time of two objects but require a reference number in class of time to compare against.
}

