% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{to_dhour}
\alias{to_dhour}
\title{Time to decimal hour converter}
\usage{
to_dhour(times, mid = F)
}
\arguments{
\item{times}{a vector of times class 'time'}

\item{mid}{whether you want to convert to 24+ time for analysis across midnight. this is off by default}
}
\description{
Returns a time in decimal hours.
}

