% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks_to_issf.R
\name{track_to_issf}
\alias{track_to_issf}
\title{Convert tracks to integrated step selection function data
Augments the tracks with a number of random generated step from the
null distribution of the \code{step_2d} and \code{delta_yaw}.}
\usage{
track_to_issf(tracks, n_sim = 100)
}
\arguments{
\item{tracks}{a \code{data.frame} with a least the columns \code{equal_time_id},
\code{step_2d} and \code{delta_yaw}.}

\item{n_sim}{Number of random alternatives}
}
\description{
Convert tracks to integrated step selection function data
Augments the tracks with a number of random generated step from the
null distribution of the \code{step_2d} and \code{delta_yaw}.
}
