% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_nearest_windmill.R
\name{update_nearest_windmill}
\alias{update_nearest_windmill}
\title{Calculate the distance to the nearest windmill}
\usage{
update_nearest_windmill(local, step = 1000)
}
\arguments{
\item{local}{the connection to the local database}

\item{step}{The number of tracks to handle in a single step.}
}
\description{
Calculate the distance to the nearest windmill
}
