\name{VariablesLDR}
\alias{VariablesLDR}
\title{Linear depolarization ratio}
\usage{VariablesLDR(Zh, Zv)}
\description{\code{VariablesLDR} returns the linear depolarization ratio
From Rinehart (1997), Eqn 10.3
}
\arguments{
  \item{Zh}{Horizontal reflectivity [mm^6/m^3]}
  \item{Zv}{Vertical reflectivity [mm^6/m^3]}
}
\value{
  \item{LDR}{ linear depolarization ratio}
}
\source{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}

R. E. Rinehart, 1997
Radar for Meteorologists
Rinehart Publishing
}
\references{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}

R. E. Rinehart, 1997
Radar for Meteorologists
Rinehart Publishing
}
\author{Jose Gama}
\keyword{programming}
