\name{VariablesRadVel}
\alias{VariablesRadVel}
\title{Radial velocity}
\usage{VariablesRadVel(f,lam)}
\description{\code{VariablesRadVel} returns the radial velocity
From Rinehart (1993), Eqn 6.6
}
\arguments{
  \item{f}{Frequency shift [Hz]}
  \item{lam}{Radar wavelength [m]}
}
\value{
  \item{Vr}{ Radial velocity [m/s]}
}
\source{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}

R. E. Rinehart, 1997
Radar for Meteorologists
Rinehart Publishing
}
\references{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}

R. E. Rinehart, 1997
Radar for Meteorologists
Rinehart Publishing
}
\author{Jose Gama}
\keyword{programming}
