\name{VariablesZDP}
\alias{VariablesZDP}
\title{Reflectivity difference}
\usage{VariablesZDP(Zh, Zv)}
\description{\code{VariablesZDP} returns the reflectivity difference
From Rinehart (1997), Eqn 10.2
}
\arguments{
  \item{Zh}{Horizontal reflectivity [mm^6/m^3]}
  \item{Zv}{Vertical reflectivity [mm^6/m^3]}
}
\value{
  \item{ZDP}{ Reflectivity difference [dB]}
}
\source{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}

R. E. Rinehart, 1997
Radar for Meteorologists
Rinehart Publishing
}
\references{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}

R. E. Rinehart, 1997
Radar for Meteorologists
Rinehart Publishing
}
\author{Jose Gama}
\keyword{programming}
