\name{VariablesZDR}
\alias{VariablesZDR}
\title{Differential reflectivity}
\usage{VariablesZDR(Zh, Zv)}
\description{\code{VariablesZDR} returns the differential reflectivity
From Rinehart (1997), Eqn 10.3 and Seliga and Bringi (1976)
}
\arguments{
  \item{Zh}{Horizontal reflectivity [mm^6/m^3]}
  \item{Zv}{Vertical reflectivity [mm^6/m^3]}
}
\value{
  \item{ZDR}{ Differential reflectivity [dB]}
}
\source{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}

R. E. Rinehart, 1997
Radar for Meteorologists
Rinehart Publishing
}
\references{
Nick Guy, 2014
PyRadarMet - Python Fundamental Calculations in Radar Meteorology
\url{https://github.com/nguy/PyRadarMet}

R. E. Rinehart, 1997
Radar for Meteorologists
Rinehart Publishing
}
\author{Jose Gama}
\keyword{programming}
