% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windmill_voxel.R
\name{windmill_voxel}
\alias{windmill_voxel}
\title{Get the number of tracks per voxels around the windmills}
\usage{
windmill_voxel(
  local,
  max_distance = 200,
  voxel_size = rep(max_distance/5, 3),
  voxel_box = matrix(c(-10, -10, 0, 10, 10, 20), nrow = 3),
  rate = 2
)
}
\arguments{
\item{local}{the connection to the local database}

\item{max_distance}{Use all equal time tracks where the minimum distance in
meters between the track and the centre is smaller than \code{max_distance}.}

\item{voxel_size}{Size in meters of the voxel.}

\item{voxel_box}{Defines the area around the centre for to set of voxels to
use.
Defined using the index number of the voxels.
Defaults to 10 voxels on both side of the centre in x and y and 20 voxels
above 0 in the z direction.}

\item{rate}{the time between two points in the output track.}
}
\description{
Get the number of tracks per voxels around the windmills
}
