% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countReads.R
\name{countReads}
\alias{countReads}
\title{quantify reads in consecutive bins}
\usage{
countReads(samplenames, gtf, fragmentLength = 150, bamFolder,
  outputDir = NA, modification = "m6A", binSize = 50,
  strandToKeep = "opposite", paired = FALSE, threads = 1)
}
\arguments{
\item{samplenames}{The names of each sample (prefix for bam files)}

\item{gtf}{The gtf format gene annotation file}

\item{fragmentLength}{The RNA fragment length (insert size of the library).}

\item{bamFolder}{Path to the folder where bam file locates}

\item{outputDir}{The directory to save output files}

\item{modification}{The middle file name used to differentiate IP sample from Input sample. This should correspond to the file name of BAM files. 
For example, modification="m6A" for BAM files named as "samplenames.m6A.bam".}

\item{binSize}{The size of consecutive bins to slice the transcripts}

\item{strandToKeep}{According to library preparation protocol, choose which strand to count. Stranded RNA library usually seq the "ooposite" strand. Small RNA library seq the "same" strand.}

\item{paired}{Logical indicating whether the input bam files are from paired end sequencing. Default is FALSE. If using paired end data, the read length will be estimated from the data and only good mate are counted.}

\item{threads}{The number of threads to use for hyperthreading}
}
