% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{diffIP}
\alias{diffIP}
\title{Perform inferential test using Poisson Random effect model in RADAR}
\usage{
\S4method{diffIP}{MeRIP.RADAR}(object, exclude = NULL, maxPsi = 100,
  fdrBy = "fdr", steps = 10, gamma = 0.5, down = 0.1)
}
\arguments{
\item{object}{The MeRIP.RADAR object}

\item{fdrBy}{The method to control for false discovery rate. The default is "qvalue", can also be "fdr".}
}
\description{
Perform inferential test using Poisson Random effect model in RADAR
}
