% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffIP.R
\name{diffIP_parallel}
\alias{diffIP_parallel}
\title{diffIP_parallel}
\usage{
diffIP_parallel(x, Covariates = NULL, plotPvalue = TRUE, p.adjBy = "fdr",
  exclude = NULL, maxPsi = 100, thread = 4)
}
\arguments{
\item{x}{The RNADMethyl data list}

\item{Covariates}{Matrix of covariates to include in the PoissonGamma test. Default is NULL.}

\item{p.adjBy}{The method used to adjust for the p value. Default is 'fdr'. Can be "bonferroni","BY","fdr".}

\item{exclude}{Names of samples to exclude (as outlier) in the test. Default is NULL}

\item{maxPsi}{The cutoff value for random effect parameter Psi estimation}

\item{thread}{The number threads to run in parallel}
}
