% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{do_feature_selection}
\alias{do_feature_selection}
\title{Perform feature selection based on different methods}
\usage{
do_feature_selection(rdr = NULL, n_features = NULL, select_by = NULL,
  method = "hcl", surv_obj = NULL, which_data = "scaled",
  corr_measure = "pearson", min_features_per_group = 5,
  thr_pca_cum_prop = 0.9, response = NULL)
}
\arguments{
\item{rdr}{A RadAR object (class \code{\link{SummarizedExperiment}}).}

\item{n_features}{(numeric) Number of features to be selected. Required.}

\item{select_by}{(character) Which criteria use to select informative radiomic features within clusters
of similar (i.e., redundant) features. It can be one of the following: "variability", "random", "concordance".}

\item{method}{(character) Which  method use to identify redundant features. It can be one of the following:
"mRMR" (minimum-redundancy-maximum-relevance),"hcl" (hierarchical clustering of correlation matrix) or "pca" (K-means applied to Principal Component Analysis).}

\item{surv_obj}{An object of class \code{\link{Surv}}. Required if select_by is "concordance".}

\item{which_data}{(character) Which data use to compute concordance index. It can be one of the following: "normal",
"scaled", "normalized".}

\item{corr_measure}{(character) Which method use to calculate correlation. It can be one of the following:
"pearson", "kendall", "spearman".}

\item{min_features_per_group}{(numeric) Minimum number of features for each cluster.}

\item{thr_pca_cum_prop}{(numeric) Threshold to select number of components based on
cumulative proportion of explained variance criterion.}

\item{response}{(numeric) A response variable, required if mRMR method is used.}
}
\value{
A list including two elements:
`rdr`: the updated (reduced) rdr (a RadAR object)
`signature`: the radiomic features included in the signature
}
\description{
This function implements different methods to perform feature selection of radiomic datasets
}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
