% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_time_track.R
\name{equal_time_track}
\alias{equal_time_track}
\title{Convert a irregular track into a track with equal time intervals}
\usage{
equal_time_track(raw_track, rate = 2)
}
\arguments{
\item{raw_track}{A single track as a data.frame with projected space-time
coordinates.
Coordinates should be in meters, the time in seconds.}

\item{rate}{the time between two points in the output track.}
}
\description{
The resulting track has
}
\details{
\itemize{
\item the space-time coordinates (\code{t}, \code{x}, \code{y} and \code{z})
\item the step length in 2D and 3D (\code{step_2d} and \code{step_3d})
\item the heading in the x-y plane (\code{yaw})
\item the heading in the plane between the z-axis and the direction in
the x-y plane (\code{pitch})
\item the change in yaw and pitch between two consecutive points (\code{delta_yaw}
and \code{delta_pitch})
}
}
