% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracks_around_centre.R
\name{get_tracks_around_centre}
\alias{get_tracks_around_centre}
\title{Extract tracks nearest to a given position}
\usage{
get_tracks_around_centre(
  local,
  centre,
  max_distance = 200,
  max_height = 200,
  max_tracks = 100,
  relevant = 1,
  rate = 2
)
}
\arguments{
\item{local}{the connection to the local database}

\item{centre}{a data frame with a single row and x and y coordinates of the
centre.
Coordinates in CRS 31370}

\item{max_distance}{Use all equal time tracks where the minimum distance in
meters between the track and the centre is smaller than \code{max_distance}.}

\item{rate}{the time between two points in the output track.}
}
\description{
Extract tracks nearest to a given position
}
