% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracks_around_windmill.R
\name{get_tracks_around_windmill}
\alias{get_tracks_around_windmill}
\title{extracts tracks nearest to windmills}
\usage{
get_tracks_around_windmill(
  local,
  max_distance = 200,
  max_height = 200,
  max_tracks = 100,
  relevant = 1,
  rate = 2,
  operator
)
}
\arguments{
\item{local}{the connection to the local database}

\item{max_distance}{Use all equal time tracks where the minimum distance in
meters between the track and the centre is smaller than \code{max_distance}.}

\item{rate}{the time between two points in the output track.}
}
\description{
extracts tracks nearest to windmills
}
