% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfToGeneModel.R
\name{gtfToGeneModel}
\alias{gtfToGeneModel}
\title{gtfToGeneModel}
\usage{
gtfToGeneModel(gtf)
}
\arguments{
\item{gtf}{gtf file to build gene model}
}
\description{
to remove ambiguous gene model and return gene model as genomic ranges object
}
