% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeLibrary.R
\name{normalizeLibrary}
\alias{normalizeLibrary}
\title{normalizeLibrary}
\usage{
normalizeLibrary(readsOut, X)
}
\arguments{
\item{readsOut}{Read counts data list from countReads() function.}

\item{X}{Study design or Grouping for the samples, should have 2 levels}
}
