% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatMap.R
\name{plotHeatMap}
\alias{plotHeatMap}
\title{plotHeatMap}
\usage{
plotHeatMap(x, covariates = NULL, topBins = "significant", label = NULL)
}
\arguments{
\item{x}{The RADAR object}

\item{covariates}{The covariates matrix, must be numerical matrix.}

\item{topBins}{The number of top bins to plot the heat map. Default is 5000.}
}
