% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/plotHeatMap.R
\docType{methods}
\name{plotHeatMap}
\alias{plotHeatMap}
\title{plot Heat Map}
\usage{
\S4method{plotHeatMap}{MeRIP.RADAR}(object, covariates = TRUE)

plotHeatMap(x, covariates = NULL, topBins = "significant",
  label = NULL)
}
\arguments{
\item{object}{The MeRIP.RADAR object.}

\item{covariates}{Logic parameter. Whether to regress out covariates (if variable has been set to have more than one column). Default is TRUE.}

\item{x}{The RADAR object}

\item{topBins}{The number of top bins to plot the heat map. Default is 5000.}

\item{covariates}{The covariates matrix, must be numerical matrix.}
}
\description{
Plot heat map of methylation variations across samples been analyzed.
}
