% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPoissonGammaMerge.R
\name{reportPoissonGammaMerge}
\alias{reportPoissonGammaMerge}
\title{reportPoissonGammaMerge}
\usage{
reportPoissonGammaMerge(x, cutoff, est = "auto")
}
\arguments{
\item{x}{The RNADMethyl data list}

\item{cutoff}{The p_value cutoff to merge the bin}

\item{est}{By default "auto", the function takes estimates from given The RNADMethyl data list. One can also pass a test estimates by assigning it to stats.}
}
