% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{test_radiomic_signature}
\alias{test_radiomic_signature}
\title{Test radiomic signature by Cox regression model}
\usage{
test_radiomic_signature(rdr = NULL, surv_obj = NULL,
  signature = NULL, which_data = "scaled")
}
\arguments{
\item{rdr}{A RadAR object (class \code{\link{SummarizedExperiment}}).}

\item{surv_obj}{An object of class \code{\link{Surv}}.}

\item{signature}{(character) vector of radiomic features defining the signature}

\item{which_data}{(character) Which data use to compute concordance index. It can be one of the following: "normal",
"scaled", "normalized".}
}
\value{
An updated (reduced) rdr (a RadAR object)
}
\description{
Test radiomic signature by Cox regression model
}
\author{
Matteo Benelli (\email{matteo.benelli@uslcentro.toscana.it})
}
